+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v3.2.0
 
-...
+### Features
+
+* Builtin Prosody: new settings to enable local C2S. For example, can be used with Matterbridge (thanks https://github.com/tytan652)
 
 ## v3.1.0